search: move lucky search out of jobutil
