+++ internal/search/lucky/feeling_lucky_search_job_test.go
@@ -1,4 +1,4 @@
-package jobutil
+package lucky
 
 import (
 	"context"
