+++ internal/search/lucky/testdata/TestCombinations/narrow_and_widen_rules.golden
