+++ internal/search/lucky/testdata/TestCombinations/only_narrow_rules.golden
