+++ internal/search/lucky/testdata/TestCombinations/only_widen_rules.golden
