+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/one_of_many_patterns_as_lang.golden
