+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/pattern_as_type.golden
