+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/pattern_as_type_multi_patterns.golden
