+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/pattern_as_type_with_expression.golden
