+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/single_pattern_as_lang.golden
