+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/trigger_unordered_patterns.golden
