+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/trigger_unquoted_rule.golden
