+++ internal/search/lucky/testdata/TestNewFeelingLuckySearchJob/type_and_lang_multi_rule.golden
