+++ internal/search/lucky/testdata/Test_patternFilter/URL_blob.golden
