+++ internal/search/lucky/testdata/Test_patternFilter/URL_pattern_as_partially_qualified_repo_filter.golden
