+++ internal/search/lucky/testdata/Test_patternFilter/looks_like_a_path.golden
