+++ internal/search/lucky/testdata/Test_patternFilter/schemaless_URL_pattern_as_repo_filter.golden
