+++ internal/search/lucky/testdata/Test_patternsAsFilters/URL_blob.golden
