+++ internal/search/lucky/testdata/Test_patternsAsFilters/URL_pattern_as_partially_qualified_repo_filter.golden
