+++ internal/search/lucky/testdata/Test_patternsAsFilters/URL_tree_path.golden
