+++ internal/search/lucky/testdata/Test_patternsAsFilters/schemaless_URL_pattern_as_repo_filter.golden
