+++ internal/search/lucky/testdata/Test_regexpPatterns/invalid_regular_expression.golden
