+++ internal/search/lucky/testdata/Test_regexpPatterns/pattern_without_enough_regexp_syntax.golden
