+++ internal/search/lucky/testdata/Test_regexpPatterns/valid_regular_expression.golden
