+++ internal/search/lucky/testdata/Test_regexpPatterns/valid_regular_expression_with_capture_group.golden
